fix file separation
