+++ src/fetch_agent/fetch_agent/communication/registry.py
@@ -15,29 +15,15 @@ from aca_protocols.station_register_protocol import (
 )
 
 from aca_protocols.property_query_protocol import (
-    PropertyQueryRequest,
-    PropertyQueryResponse,
     PropertyData,
 )
 
 from aca_protocols.acs_registry_id import acs_id
 
-protocol = Protocol()
+stationRegisterProtocol = Protocol()
 
 
-@protocol.on_message(model=PropertyQueryRequest, replies=PropertyQueryResponse)
-async def on_query_properties(ctx: Context, sender: str, request: PropertyQueryRequest):
-    ctx.logger.info(f"{sender} requested params")
-
-    properties = ctx.storage.get("properties")
-    properties = PropertyData.from_json(properties)
-
-    response = PropertyQueryResponse(properties=properties)
-
-    await ctx.send(sender, response)
-
-
-@protocol.on_message(model=StationRegisterResponse)
+@stationRegisterProtocol.on_message(model=StationRegisterResponse)
 async def on_is_registered(ctx: Context, sender: str, request: StationRegisterResponse):
     ctx.logger.info(f"got registered by: {sender}; TTL: {request.ttl}")
     ctx.storage.set(
